# fix(es/transforms/compat): Improve performance