--- a/ecmascript/transforms/compat/src/es2015/block_scoped_fn.rs
+++ b/ecmascript/transforms/compat/src/es2015/block_scoped_fn.rs
@@ -1,6 +1,12 @@
 use swc_common::{Spanned, DUMMY_SP};
 use swc_ecma_ast::*;
+use swc_ecma_transforms_base::perf::Check;
+use swc_ecma_transforms_macros::fast_path;
 use swc_ecma_utils::UsageFinder;
+use swc_ecma_visit::noop_visit_type;
+use swc_ecma_visit::Node;
+use swc_ecma_visit::Visit;
+use swc_ecma_visit::VisitWith;
 use swc_ecma_visit::{noop_fold_type, Fold, FoldWith};
 
 pub fn block_scoped_functions() -> impl Fold {
@@ -10,6 +16,7 @@ pub fn block_scoped_functions() -> impl Fold {
 #[derive(Clone, Copy)]
 struct BlockScopedFns;
 
+#[fast_path(BlockScopedFnFinder)]
 impl Fold for BlockScopedFns {
     noop_fold_type!();
 
@@ -58,6 +65,32 @@ impl Fold for BlockScopedFns {
         stmts.append(&mut extra_stmts);
 
         stmts
+    }
+}
+
+#[derive(Default)]
+struct BlockScopedFnFinder {
+    found: bool,
+}
+
+impl Visit for BlockScopedFnFinder {
+    noop_visit_type!();
+
+    fn visit_stmts(&mut self, stmts: &[Stmt], _: &dyn Node) {
+        for n in stmts {
+            n.visit_with(&Invalid { span: DUMMY_SP }, self);
+        }
+
+        self.found |= stmts.iter().any(|stmt| match stmt {
+            Stmt::Decl(Decl::Fn(..)) => true,
+            _ => false,
+        });
+    }
+}
+
+impl Check for BlockScopedFnFinder {
+    fn should_handle(&self) -> bool {
+        self.found
     }
 }