--- a/ecmascript/transforms/compat/src/es2015/sticky_regex.rs
+++ b/ecmascript/transforms/compat/src/es2015/sticky_regex.rs
@@ -1,8 +1,13 @@
 use swc_atoms::JsWord;
 use swc_common::DUMMY_SP;
 use swc_ecma_ast::*;
+use swc_ecma_transforms_base::perf::Check;
+use swc_ecma_transforms_macros::fast_path;
 use swc_ecma_utils::quote_ident;
 use swc_ecma_utils::ExprFactory;
+use swc_ecma_visit::noop_visit_type;
+use swc_ecma_visit::Node;
+use swc_ecma_visit::Visit;
 use swc_ecma_visit::{noop_fold_type, Fold, FoldWith};
 
 /// Compile ES2015 sticky regex to an ES5 RegExp constructor
@@ -26,6 +31,7 @@ pub fn sticky_regex() -> impl 'static + Fold {
 #[derive(Clone, Copy)]
 struct StickyRegex;
 
+#[fast_path(StickyRegexVisitor)]
 impl Fold for StickyRegex {
     noop_fold_type!();
 
@@ -58,6 +64,25 @@ impl Fold for StickyRegex {
             }
             _ => e,
         }
+    }
+}
+
+#[derive(Default)]
+struct StickyRegexVisitor {
+    found: bool,
+}
+
+impl Visit for StickyRegexVisitor {
+    noop_visit_type!();
+
+    fn visit_regex(&mut self, n: &Regex, _: &dyn Node) {
+        self.found |= n.flags.contains('y');
+    }
+}
+
+impl Check for StickyRegexVisitor {
+    fn should_handle(&self) -> bool {
+        self.found
     }
 }