--- a/ecmascript/transforms/compat/src/es2015/typeof_symbol.rs
+++ b/ecmascript/transforms/compat/src/es2015/typeof_symbol.rs
@@ -1,7 +1,8 @@
 use swc_atoms::js_word;
-use swc_common::DUMMY_SP;
 use swc_ecma_ast::*;
 use swc_ecma_transforms_base::helper;
+use swc_ecma_transforms_base::perf::Check;
+use swc_ecma_transforms_macros::fast_path;
 use swc_ecma_utils::ExprFactory;
 use swc_ecma_visit::noop_visit_type;
 use swc_ecma_visit::{noop_fold_type, Fold, FoldWith, Node, Visit, VisitWith};
@@ -13,15 +14,11 @@ pub fn typeof_symbol() -> impl Fold {
 #[derive(Clone)]
 struct TypeOfSymbol;
 
+#[fast_path(TypeOfFinder)]
 impl Fold for TypeOfSymbol {
     noop_fold_type!();
 
     fn fold_expr(&mut self, expr: Expr) -> Expr {
-        // fast path
-        if !should_work(&expr) {
-            return expr;
-        }
-
         let expr = expr.fold_children_with(self);
 
         match expr {
@@ -71,22 +68,27 @@ impl Fold for TypeOfSymbol {
     }
 }
 
-fn should_work(node: &Expr) -> bool {
-    struct Visitor {
-        found: bool,
-    }
-    impl Visit for Visitor {
-        noop_visit_type!();
+#[derive(Default)]
+struct TypeOfFinder {
+    found: bool,
+}
 
-        fn visit_unary_expr(&mut self, e: &UnaryExpr, _: &dyn Node) {
-            if e.op == op!("typeof") {
-                self.found = true
-            }
+impl Visit for TypeOfFinder {
+    noop_visit_type!();
+
+    fn visit_unary_expr(&mut self, e: &UnaryExpr, _: &dyn Node) {
+        e.visit_children_with(self);
+
+        if e.op == op!("typeof") {
+            self.found = true
         }
     }
-    let mut v = Visitor { found: false };
-    node.visit_with(&Invalid { span: DUMMY_SP } as _, &mut v);
-    v.found
+}
+
+impl Check for TypeOfFinder {
+    fn should_handle(&self) -> bool {
+        self.found
+    }
 }
 
 fn is_non_symbol_literal(e: &Expr) -> bool {