# Tasks


swc_ecma_transforms_compat:
 - [x] `classes`: Fast-path.
 - [x] `destructuring`: Fast-path.
 - [x] `sticky_regex`: Fast-path.
 - [x] `spread`: Fast-path.
 - [x] `shorthand`: Fast-path.
 - [x] `typeof_symbol`: Fast-path.
 - [x] `block_scoped_functions`: Fast path.
 - [x] `duplicate_keys`: Use fxhash.
 - [x] `instance_of`: Fast path.
 - [x] `reserved_words`: Make string comparison efficient.


## Initial

```
test base                          ... bench:     111,254 ns/iter (+/- 13,235) = 97 MB/s
test common_reserved_word          ... bench:     121,888 ns/iter (+/- 16,508) = 89 MB/s
test common_typescript             ... bench:     269,342 ns/iter (+/- 30,072) = 40 MB/s
test es2015                        ... bench:   1,766,677 ns/iter (+/- 125,103) = 6 MB/s
test es2015_arrow                  ... bench:     162,651 ns/iter (+/- 19,908) = 66 MB/s
test es2015_block_scoped_fn        ... bench:     203,907 ns/iter (+/- 25,950) = 53 MB/s
test es2015_block_scoping          ... bench:     190,847 ns/iter (+/- 22,505) = 57 MB/s
test es2015_classes                ... bench:     451,819 ns/iter (+/- 48,700) = 24 MB/s
test es2015_computed_props         ... bench:     119,147 ns/iter (+/- 18,644) = 91 MB/s
test es2015_destructuring          ... bench:     356,579 ns/iter (+/- 38,148) = 30 MB/s
test es2015_duplicate_keys         ... bench:     199,743 ns/iter (+/- 23,338) = 54 MB/s
test es2015_fn_name                ... bench:     192,128 ns/iter (+/- 19,673) = 56 MB/s
test es2015_for_of                 ... bench:     118,548 ns/iter (+/- 17,155) = 91 MB/s
test es2015_instanceof             ... bench:     133,513 ns/iter (+/- 13,504) = 81 MB/s
test es2015_parameters             ... bench:     199,903 ns/iter (+/- 21,725) = 54 MB/s
test es2015_shorthand_property     ... bench:     179,865 ns/iter (+/- 20,189) = 60 MB/s
test es2015_spread                 ... bench:     187,041 ns/iter (+/- 23,023) = 58 MB/s
test es2015_sticky_regex           ... bench:     188,447 ns/iter (+/- 25,198) = 57 MB/s
test es2015_typeof_symbol          ... bench:     135,419 ns/iter (+/- 19,231) = 80 MB/s
test es2016                        ... bench:     117,282 ns/iter (+/- 14,795) = 92 MB/s
test es2016_exponentation          ... bench:     117,254 ns/iter (+/- 17,355) = 92 MB/s
test es2017                        ... bench:     113,073 ns/iter (+/- 12,918) = 96 MB/s
test es2017_async_to_generator     ... bench:     115,553 ns/iter (+/- 15,809) = 94 MB/s
test es2018                        ... bench:     130,281 ns/iter (+/- 13,567) = 83 MB/s
test es2018_object_rest_spread     ... bench:     123,031 ns/iter (+/- 11,428) = 88 MB/s
test es2018_optional_catch_binding ... bench:     114,980 ns/iter (+/- 14,354) = 94 MB/s
test es2020                        ... bench:     337,954 ns/iter (+/- 45,369) = 32 MB/s
test es2020_class_properties       ... bench:     247,589 ns/iter (+/- 36,081) = 43 MB/s
test es2020_nullish_coalescing     ... bench:     113,015 ns/iter (+/- 9,730) = 96 MB/s
test es2020_optional_chaining      ... bench:     172,059 ns/iter (+/- 16,100) = 63 MB/s
test es3                           ... bench:     368,856 ns/iter (+/- 36,758) = 29 MB/s
test full_es2016                   ... bench:     381,481 ns/iter (+/- 40,132) = 28 MB/s
test full_es2017                   ... bench:     375,360 ns/iter (+/- 48,894) = 29 MB/s
test full_es2018                   ... bench:     368,916 ns/iter (+/- 35,474) = 29 MB/s
```


## After adding a fast path to `classes`

```
test base                          ... bench:     119,123 ns/iter (+/- 11,349) = 91 MB/s
test common_reserved_word          ... bench:     128,702 ns/iter (+/- 17,183) = 84 MB/s
test common_typescript             ... bench:     271,088 ns/iter (+/- 30,242) = 40 MB/s
test es2015                        ... bench:   1,720,555 ns/iter (+/- 175,216) = 6 MB/s
test es2015_arrow                  ... bench:     164,435 ns/iter (+/- 24,196) = 66 MB/s
test es2015_block_scoped_fn        ... bench:     200,050 ns/iter (+/- 20,588) = 54 MB/s
test es2015_block_scoping          ... bench:     192,415 ns/iter (+/- 27,862) = 56 MB/s
test es2015_classes                ... bench:     387,521 ns/iter (+/- 34,718) = 28 MB/s
test es2015_computed_props         ... bench:     117,924 ns/iter (+/- 16,929) = 92 MB/s
test es2015_destructuring          ... bench:     351,689 ns/iter (+/- 49,016) = 30 MB/s
test es2015_duplicate_keys         ... bench:     194,642 ns/iter (+/- 20,593) = 55 MB/s
test es2015_fn_name                ... bench:     196,668 ns/iter (+/- 26,409) = 55 MB/s
test es2015_for_of                 ... bench:     114,456 ns/iter (+/- 15,472) = 95 MB/s
test es2015_instanceof             ... bench:     132,189 ns/iter (+/- 15,908) = 82 MB/s
test es2015_parameters             ... bench:     204,334 ns/iter (+/- 24,661) = 53 MB/s
test es2015_shorthand_property     ... bench:     184,448 ns/iter (+/- 22,869) = 59 MB/s
test es2015_spread                 ... bench:     189,732 ns/iter (+/- 25,938) = 57 MB/s
test es2015_sticky_regex           ... bench:     190,018 ns/iter (+/- 27,346) = 57 MB/s
test es2015_typeof_symbol          ... bench:     139,368 ns/iter (+/- 19,035) = 78 MB/s
test es2016                        ... bench:     112,390 ns/iter (+/- 14,009) = 96 MB/s
test es2016_exponentation          ... bench:     114,322 ns/iter (+/- 12,903) = 95 MB/s
test es2017                        ... bench:     115,519 ns/iter (+/- 14,461) = 94 MB/s
test es2017_async_to_generator     ... bench:     113,843 ns/iter (+/- 15,975) = 95 MB/s
test es2018                        ... bench:     128,982 ns/iter (+/- 11,161) = 84 MB/s
test es2018_object_rest_spread     ... bench:     121,391 ns/iter (+/- 10,947) = 89 MB/s
test es2018_optional_catch_binding ... bench:     116,146 ns/iter (+/- 13,819) = 93 MB/s
test es2020                        ... bench:     336,522 ns/iter (+/- 33,591) = 32 MB/s
test es2020_class_properties       ... bench:     242,389 ns/iter (+/- 24,465) = 44 MB/s
test es2020_nullish_coalescing     ... bench:     114,119 ns/iter (+/- 12,118) = 95 MB/s
test es2020_optional_chaining      ... bench:     168,943 ns/iter (+/- 21,376) = 64 MB/s
test es3                           ... bench:     359,321 ns/iter (+/- 37,489) = 30 MB/s
test full_es2016                   ... bench:     380,527 ns/iter (+/- 37,205) = 28 MB/s
test full_es2017                   ... bench:     375,781 ns/iter (+/- 36,744) = 28 MB/s
test full_es2018                   ... bench:     363,945 ns/iter (+/- 31,571) = 29 MB/s
```


## After adding a fast path to `destructuring`

```
running 34 tests
test base                          ... bench:     110,978 ns/iter (+/- 12,179) = 98 MB/s
test common_reserved_word          ... bench:     124,226 ns/iter (+/- 9,441) = 87 MB/s
test common_typescript             ... bench:     272,462 ns/iter (+/- 26,837) = 39 MB/s
test es2015                        ... bench:   1,610,694 ns/iter (+/- 164,953) = 6 MB/s
test es2015_arrow                  ... bench:     169,458 ns/iter (+/- 19,944) = 64 MB/s
test es2015_block_scoped_fn        ... bench:     200,063 ns/iter (+/- 30,689) = 54 MB/s
test es2015_block_scoping          ... bench:     193,180 ns/iter (+/- 26,777) = 56 MB/s
test es2015_classes                ... bench:     387,143 ns/iter (+/- 44,264) = 28 MB/s
test es2015_computed_props         ... bench:     117,002 ns/iter (+/- 16,426) = 93 MB/s
test es2015_destructuring          ... bench:     217,475 ns/iter (+/- 19,423) = 50 MB/s
test es2015_duplicate_keys         ... bench:     197,313 ns/iter (+/- 25,820) = 55 MB/s
test es2015_fn_name                ... bench:     192,425 ns/iter (+/- 19,599) = 56 MB/s
test es2015_for_of                 ... bench:     114,706 ns/iter (+/- 13,315) = 94 MB/s
test es2015_instanceof             ... bench:     132,504 ns/iter (+/- 16,860) = 82 MB/s
test es2015_parameters             ... bench:     207,415 ns/iter (+/- 32,322) = 52 MB/s
test es2015_shorthand_property     ... bench:     184,097 ns/iter (+/- 28,718) = 59 MB/s
test es2015_spread                 ... bench:     185,806 ns/iter (+/- 18,058) = 58 MB/s
test es2015_sticky_regex           ... bench:     185,061 ns/iter (+/- 18,635) = 58 MB/s
test es2015_typeof_symbol          ... bench:     132,213 ns/iter (+/- 14,799) = 82 MB/s
test es2016                        ... bench:     112,725 ns/iter (+/- 15,487) = 96 MB/s
test es2016_exponentation          ... bench:     116,831 ns/iter (+/- 19,807) = 93 MB/s
test es2017                        ... bench:     115,354 ns/iter (+/- 13,631) = 94 MB/s
test es2017_async_to_generator     ... bench:     112,624 ns/iter (+/- 10,224) = 96 MB/s
test es2018                        ... bench:     135,870 ns/iter (+/- 16,089) = 80 MB/s
test es2018_object_rest_spread     ... bench:     123,835 ns/iter (+/- 16,963) = 87 MB/s
test es2018_optional_catch_binding ... bench:     116,783 ns/iter (+/- 17,807) = 93 MB/s
test es2020                        ... bench:     341,832 ns/iter (+/- 38,694) = 31 MB/s
test es2020_class_properties       ... bench:     245,007 ns/iter (+/- 21,921) = 44 MB/s
test es2020_nullish_coalescing     ... bench:     113,899 ns/iter (+/- 10,124) = 95 MB/s
test es2020_optional_chaining      ... bench:     166,763 ns/iter (+/- 16,423) = 65 MB/s
test es3                           ... bench:     361,489 ns/iter (+/- 44,733) = 30 MB/s
test full_es2016                   ... bench:     376,102 ns/iter (+/- 37,820) = 28 MB/s
test full_es2017                   ... bench:     372,284 ns/iter (+/- 41,281) = 29 MB/s
test full_es2018                   ... bench:     368,988 ns/iter (+/- 35,504) = 29 MB/s
```

## After adding a fast path to `sticky_regex`

```
test es2015_sticky_regex           ... bench:     109,634 ns/iter (+/- 19,760) = 99 MB/s
```


## After adding a fast path to `spread`

```
running 34 tests
test base                          ... bench:     110,467 ns/iter (+/- 14,280) = 98 MB/s
test common_reserved_word          ... bench:     120,610 ns/iter (+/- 21,053) = 90 MB/s
test common_typescript             ... bench:     262,416 ns/iter (+/- 21,867) = 41 MB/s
test es2015                        ... bench:   1,415,561 ns/iter (+/- 173,442) = 7 MB/s
test es2015_arrow                  ... bench:     159,041 ns/iter (+/- 25,010) = 68 MB/s
test es2015_block_scoped_fn        ... bench:     191,183 ns/iter (+/- 21,420) = 56 MB/s
test es2015_block_scoping          ... bench:     185,045 ns/iter (+/- 17,661) = 58 MB/s
test es2015_classes                ... bench:     376,451 ns/iter (+/- 30,514) = 28 MB/s
test es2015_computed_props         ... bench:     116,755 ns/iter (+/- 12,743) = 93 MB/s
test es2015_destructuring          ... bench:     207,133 ns/iter (+/- 19,530) = 52 MB/s
test es2015_duplicate_keys         ... bench:     190,161 ns/iter (+/- 19,727) = 57 MB/s
test es2015_fn_name                ... bench:     185,233 ns/iter (+/- 28,422) = 58 MB/s
test es2015_for_of                 ... bench:     115,010 ns/iter (+/- 18,323) = 94 MB/s
test es2015_instanceof             ... bench:     135,199 ns/iter (+/- 17,909) = 80 MB/s
test es2015_parameters             ... bench:     202,255 ns/iter (+/- 23,682) = 53 MB/s
test es2015_shorthand_property     ... bench:     180,277 ns/iter (+/- 25,541) = 60 MB/s
test es2015_spread                 ... bench:     114,339 ns/iter (+/- 19,182) = 95 MB/s
test es2015_sticky_regex           ... bench:     115,323 ns/iter (+/- 13,302) = 94 MB/s
test es2015_typeof_symbol          ... bench:     131,391 ns/iter (+/- 17,722) = 82 MB/s
test es2016                        ... bench:     114,611 ns/iter (+/- 16,855) = 95 MB/s
test es2016_exponentation          ... bench:     113,137 ns/iter (+/- 16,839) = 96 MB/s
test es2017                        ... bench:     114,441 ns/iter (+/- 15,281) = 95 MB/s
test es2017_async_to_generator     ... bench:     114,061 ns/iter (+/- 12,425) = 95 MB/s
test es2018                        ... bench:     127,693 ns/iter (+/- 9,025) = 85 MB/s
test es2018_object_rest_spread     ... bench:     120,077 ns/iter (+/- 15,942) = 90 MB/s
test es2018_optional_catch_binding ... bench:     114,473 ns/iter (+/- 15,175) = 95 MB/s
test es2020                        ... bench:     332,852 ns/iter (+/- 30,172) = 32 MB/s
test es2020_class_properties       ... bench:     240,001 ns/iter (+/- 24,858) = 45 MB/s
test es2020_nullish_coalescing     ... bench:     114,193 ns/iter (+/- 17,286) = 95 MB/s
test es2020_optional_chaining      ... bench:     169,462 ns/iter (+/- 20,051) = 64 MB/s
test es3                           ... bench:     361,264 ns/iter (+/- 35,778) = 30 MB/s
test full_es2016                   ... bench:     373,841 ns/iter (+/- 45,641) = 29 MB/s
test full_es2017                   ... bench:     365,681 ns/iter (+/- 47,326) = 29 MB/s
test full_es2018                   ... bench:     362,914 ns/iter (+/- 43,774) = 30 MB/s
```


## After adding a fast-path to `shorthand`


```
test base                          ... bench:     109,045 ns/iter (+/- 19,984) = 99 MB/s
test common_reserved_word          ... bench:     123,454 ns/iter (+/- 13,924) = 88 MB/s
test common_typescript             ... bench:     265,849 ns/iter (+/- 22,810) = 40 MB/s
test es2015                        ... bench:   1,329,596 ns/iter (+/- 75,531) = 8 MB/s
test es2015_arrow                  ... bench:     161,227 ns/iter (+/- 25,498) = 67 MB/s
test es2015_block_scoped_fn        ... bench:     196,059 ns/iter (+/- 27,339) = 55 MB/s
test es2015_block_scoping          ... bench:     186,595 ns/iter (+/- 26,355) = 58 MB/s
test es2015_classes                ... bench:     377,713 ns/iter (+/- 33,404) = 28 MB/s
test es2015_computed_props         ... bench:     112,094 ns/iter (+/- 16,380) = 97 MB/s
test es2015_destructuring          ... bench:     216,431 ns/iter (+/- 26,043) = 50 MB/s
test es2015_duplicate_keys         ... bench:     188,096 ns/iter (+/- 26,518) = 57 MB/s
test es2015_fn_name                ... bench:     187,868 ns/iter (+/- 24,676) = 57 MB/s
test es2015_for_of                 ... bench:     115,231 ns/iter (+/- 20,022) = 94 MB/s
test es2015_instanceof             ... bench:     128,972 ns/iter (+/- 12,909) = 84 MB/s
test es2015_parameters             ... bench:     196,952 ns/iter (+/- 22,703) = 55 MB/s
test es2015_shorthand_property     ... bench:     114,210 ns/iter (+/- 15,475) = 95 MB/s
test es2015_spread                 ... bench:     111,460 ns/iter (+/- 11,789) = 97 MB/s
test es2015_sticky_regex           ... bench:     116,540 ns/iter (+/- 17,669) = 93 MB/s
test es2015_typeof_symbol          ... bench:     129,232 ns/iter (+/- 10,847) = 84 MB/s
test es2016                        ... bench:     114,294 ns/iter (+/- 13,265) = 95 MB/s
test es2016_exponentation          ... bench:     113,178 ns/iter (+/- 15,223) = 96 MB/s
test es2017                        ... bench:     112,647 ns/iter (+/- 8,948) = 96 MB/s
test es2017_async_to_generator     ... bench:     110,528 ns/iter (+/- 9,492) = 98 MB/s
test es2018                        ... bench:     129,380 ns/iter (+/- 20,162) = 84 MB/s
test es2018_object_rest_spread     ... bench:     121,343 ns/iter (+/- 16,102) = 89 MB/s
test es2018_optional_catch_binding ... bench:     113,886 ns/iter (+/- 9,513) = 95 MB/s
test es2020                        ... bench:     337,599 ns/iter (+/- 42,616) = 32 MB/s
test es2020_class_properties       ... bench:     241,041 ns/iter (+/- 32,503) = 45 MB/s
test es2020_nullish_coalescing     ... bench:     118,513 ns/iter (+/- 21,243) = 91 MB/s
test es2020_optional_chaining      ... bench:     176,100 ns/iter (+/- 29,701) = 61 MB/s
test es3                           ... bench:     369,626 ns/iter (+/- 63,000) = 29 MB/s
test full_es2016                   ... bench:     382,746 ns/iter (+/- 56,262) = 28 MB/s
test full_es2017                   ... bench:     377,044 ns/iter (+/- 55,082) = 28 MB/s
test full_es2018                   ... bench:     365,537 ns/iter (+/- 49,721) = 29 MB/s
```


### After adding fast-paths to `typeof_symbol` and `block_scoped_functions`


```
test base                          ... bench:     111,699 ns/iter (+/- 22,995) = 97 MB/s
test common_reserved_word          ... bench:     124,373 ns/iter (+/- 17,009) = 87 MB/s
test common_typescript             ... bench:     262,763 ns/iter (+/- 28,433) = 41 MB/s
test es2015                        ... bench:   1,278,111 ns/iter (+/- 167,443) = 8 MB/s
test es2015_arrow                  ... bench:     165,562 ns/iter (+/- 17,693) = 65 MB/s
test es2015_block_scoped_fn        ... bench:     126,581 ns/iter (+/- 13,807) = 86 MB/s
test es2015_block_scoping          ... bench:     192,125 ns/iter (+/- 29,475) = 56 MB/s
test es2015_classes                ... bench:     373,021 ns/iter (+/- 45,325) = 29 MB/s
test es2015_computed_props         ... bench:     114,298 ns/iter (+/- 20,493) = 95 MB/s
test es2015_destructuring          ... bench:     215,121 ns/iter (+/- 25,235) = 50 MB/s
test es2015_duplicate_keys         ... bench:     194,354 ns/iter (+/- 18,901) = 56 MB/s
test es2015_fn_name                ... bench:     184,155 ns/iter (+/- 17,998) = 59 MB/s
test es2015_for_of                 ... bench:     110,387 ns/iter (+/- 10,112) = 98 MB/s
test es2015_instanceof             ... bench:     132,182 ns/iter (+/- 17,539) = 82 MB/s
test es2015_parameters             ... bench:     209,194 ns/iter (+/- 23,297) = 52 MB/s
test es2015_shorthand_property     ... bench:     113,278 ns/iter (+/- 14,347) = 96 MB/s
test es2015_spread                 ... bench:     111,437 ns/iter (+/- 10,849) = 97 MB/s
test es2015_sticky_regex           ... bench:     115,240 ns/iter (+/- 16,694) = 94 MB/s
test es2015_typeof_symbol          ... bench:     132,923 ns/iter (+/- 13,064) = 81 MB/s
test es2016                        ... bench:     114,351 ns/iter (+/- 15,971) = 95 MB/s
test es2016_exponentation          ... bench:     115,622 ns/iter (+/- 16,725) = 94 MB/s
test es2017                        ... bench:     112,396 ns/iter (+/- 14,309) = 96 MB/s
test es2017_async_to_generator     ... bench:     113,016 ns/iter (+/- 15,287) = 96 MB/s
test es2018                        ... bench:     129,416 ns/iter (+/- 19,047) = 84 MB/s
test es2018_object_rest_spread     ... bench:     123,075 ns/iter (+/- 16,372) = 88 MB/s
test es2018_optional_catch_binding ... bench:     117,407 ns/iter (+/- 16,016) = 92 MB/s
test es2020                        ... bench:     340,657 ns/iter (+/- 43,035) = 31 MB/s
test es2020_class_properties       ... bench:     236,191 ns/iter (+/- 20,927) = 46 MB/s
test es2020_nullish_coalescing     ... bench:     114,683 ns/iter (+/- 13,315) = 94 MB/s
test es2020_optional_chaining      ... bench:     165,987 ns/iter (+/- 19,508) = 65 MB/s
test es3                           ... bench:     359,162 ns/iter (+/- 50,442) = 30 MB/s
test full_es2016                   ... bench:     366,948 ns/iter (+/- 43,308) = 29 MB/s
test full_es2017                   ... bench:     372,805 ns/iter (+/- 42,187) = 29 MB/s
test full_es2018                   ... bench:     368,196 ns/iter (+/- 40,631) = 29 MB/s
```


## After adding fast path to `instanceof`


```
test base                          ... bench:     106,474 ns/iter (+/- 9,747) = 102 MB/s
test common_reserved_word          ... bench:     125,977 ns/iter (+/- 19,124) = 86 MB/s
test common_typescript             ... bench:     261,931 ns/iter (+/- 22,709) = 41 MB/s
test es2015                        ... bench:   1,260,394 ns/iter (+/- 111,879) = 8 MB/s
test es2015_arrow                  ... bench:     159,833 ns/iter (+/- 18,372) = 68 MB/s
test es2015_block_scoped_fn        ... bench:     125,100 ns/iter (+/- 11,913) = 87 MB/s
…